requirements.txt
streamlit
requests
google-generativeai
pytz
PyPDF2
pdfplumber
beautifulsoup4
lxml
textblob
feedparser
newsapi-python
python-dotenv